smoke/requirements.txt
behave
pyshould
requests
kubernetes
urllib3 == 1.25.9
pyfiglet
clint
python-jenkins

